Splitting the TypeScript for the Dashboard into separate files to make it easier to work with
